requirements.txt
requests
Pillow
pynput
Flask
pycryptodome
cryptography
pyinstaller